STYLE: Split main window class to allow re-use in Slicer custom application
This commit refactors the class "qSlicerAppMainWindow" and adds
the class "qSlicerMainWindow" to QTApp library.

This allows to use a standard derivation approach in Slicer custom
application.

The following user interface customizations are done in qSlicerAppMainWindow:

(1) Setting of WindowsTitle and WindowIcon

(2) Setting of the logo associated with the module panel

(3) Customization of the help menu adding the following actions:
* helpKeyboardShortcutsAction
* helpInterfaceDocumentationAction
* helpBrowseTutorialsAction
* helpSlicerPublicationsAction
* helpVisualBlogAction
* helpReportBugOrFeatureRequestAction
* helpAboutSlicerAppAction


Tested-by: Pablo Hernandez <pablo.hernandez@kitware.com>
